requirements_unetpp.txt
# U-Net++ Training Requirements
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0

# Data handling
numpy>=1.23.0
pillow>=9.0.0
pandas>=1.5.0

# Metrics and visualization
scikit-learn>=1.2.0
matplotlib>=3.5.0
tqdm>=4.64.0

# Utilities
opencv-python>=4.7.0
